middleware/requirements.txt
# Core dependencies
# Note: Engine uses external llama-server.exe via subprocess.
# No python binding required.

# HTTP client for server mode
requests>=2.28.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0

# EPUB handling
lxml>=4.9.0
beautifulsoup4>=4.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0.0

# Traditional Chinese Support
# Replaced opencc with opencc-python-reimplemented to avoid C++ build issues on Windows
opencc-python-reimplemented>=0.1.7

# Hardware Monitoring
pynvml>=11.0.0

# Term Extraction (Japanese Tokenizer)
# fugashi: MeCab wrapper for Japanese morphological analysis
# unidic-lite: Lightweight dictionary (~248MB, required)
fugashi>=1.3.0
unidic-lite>=1.0.8
